NFER-66 chore: extract a lockout persistence layer location from a storage implementation into `LockoutStorageAbstract` NFER-66 chore: inject a key-value persistence ID into `KvLockoutStorage` during the service installation
