+++ models/classes/oauth/lockout/storage/LockoutStorageAbstract.php
@@ -15,22 +15,35 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2020 (original work) (update and modification) Open Assessment Technologies SA (under the project TAO-PRODUCT)
+ * Copyright (c) 2020 (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
 
 namespace oat\tao\model\oauth\lockout\storage;
 
-class KvLockoutStorage extends LockoutStorageAbstract
+use common_persistence_Persistence as Persistence;
+use oat\generis\persistence\PersistenceManager;
+use oat\oatbox\service\ConfigurableService;
+
+abstract class LockoutStorageAbstract extends ConfigurableService implements LockoutStorageInterface
 {
-    public function store(string $ip, int $ttl = 0)
+    /** @var Persistence */
+    private $persistence;
+
+    public function getPersistenceId(): string
     {
-        return true;
+        return $this->getOption(self::OPTION_PERSISTENCE);
     }
 
-    public function getFailedAttempts(string $ip, int $timeout): int
+    protected function getPersistence(): Persistence
     {
-        return 0;
+        if ($this->persistence === null) {
+            $this->persistence = $this->getServiceLocator()
+                ->get(PersistenceManager::SERVICE_ID)
+                ->getPersistenceById($this->getPersistenceId());
+        }
+
+        return $this->persistence;
     }
 }
