+++ scripts/install/SetUpOAuthLockoutService.php
@@ -29,7 +29,6 @@ use oat\tao\model\oauth\lockout\IPLockout;
 use oat\tao\model\oauth\lockout\LockoutInterface;
 use oat\tao\model\oauth\lockout\NoLockout;
 use oat\tao\model\oauth\lockout\storage\KvLockoutStorage;
-use oat\tao\model\oauth\lockout\storage\LockoutStorageInterface;
 use oat\tao\model\oauth\lockout\storage\RdsLockoutStorage;
 use oat\tao\model\oauth\OauthService;
 
@@ -154,7 +153,7 @@ class SetUpOAuthLockoutService extends ScriptAction
         $options[IPLockout::OPTION_IP_FACTORY] = new IPFactory();
         switch ($this->getOption(self::OPT_STORAGE)) {
             case self::STORAGE_KV:
-                $options[IPLockout::OPTION_LOCKOUT_STORAGE] = new KvLockoutStorage();
+                $options[IPLockout::OPTION_LOCKOUT_STORAGE] = new KvLockoutStorage([RdsLockoutStorage::OPTION_PERSISTENCE => 'default_kv']);
                 break;
             case self::STORAGE_RDS:
                 $options[IPLockout::OPTION_LOCKOUT_STORAGE] = new RdsLockoutStorage([RdsLockoutStorage::OPTION_PERSISTENCE => 'default']);
